% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIntoMemory.R
\name{loadIntoMemory}
\alias{loadIntoMemory}
\title{Load a TileDB matrix into memory}
\usage{
loadIntoMemory(x, num.threads = 1)
}
\arguments{
\item{x}{A \pkg{TileDBArray}-derived matrix or seed object.}

\item{num.threads}{Integer scalar specifying the number of threads to use for loading.}
}
\value{
An external pointer that can be used in \pkg{tatami}-based functions.
}
\description{
Load a TileDB-backed matrix into memory as an external pointer to a \pkg{tatami}-compatible representation.
This differs from the (default) behavior of \code{\link{initializeCpp}}, which only loads slices of the matrix on request.
}
\examples{
library(TileDBArray)
y <- matrix(runif(1000), ncol=20, nrow=50)
z <- as(y, "TileDBArray")
ptr <- loadIntoMemory(z)

}
\author{
Aaron Lun
}
